refactor(types): restructure Organization type and update its imports
