+++ src/app/qrcode/qr-url/page.tsx
@@ -12,9 +12,8 @@ import AuthContext from '@/contexts/authContext';
 import { useAuthRouter, useScreenSize } from '@/hooks';
 import { meService, categoryService, organizationService } from '@/services';
 
-import { CategoryColor } from '@/types';
+import { CategoryColor, Organization } from '@/types';
 import Category from '@/types/category-type';
-import Organization from '@/types/organization-type';
 import ScreenSize from '@/types/screen-size-enum';
 import ShortenInputFieldEnum from '@/types/shorten-input-field-enum';
 import ShortenInputFieldType from '@/types/shorten-input-field-type';
