+++ src/app/qrcode/qr-wifi/page.tsx
@@ -14,10 +14,9 @@ import { useAuthRouter, useScreenSize } from '@/hooks';
 import { meService, categoryService, organizationService } from '@/services';
 import { getIcon } from '@/utils/common';
 
-import { CategoryColor } from '@/types';
+import { CategoryColor, Organization } from '@/types';
 import Category from '@/types/category-type';
 import Icon from '@/types/icon-enum';
-import Organization from '@/types/organization-type';
 import ScreenSize from '@/types/screen-size-enum';
 import ShortenInputFieldEnum from '@/types/shorten-input-field-enum';
 import ShortenInputFieldType from '@/types/shorten-input-field-type';
