+++ src/services/organization.ts
@@ -1,20 +1,14 @@
-import Category from '@/types/category-type';
-import Organization from '@/types/organization-type';
-import Url from '@/types/url-type';
-
 import api from './api';
 
-async function getCategoryByOrganizationId(
-  organizationId: Organization['_id'],
-  page: number = 1,
-  limit: number = 10,
-) {
+import { Organization } from '@/types';
+import Category from '@/types/category-type';
+import Url from '@/types/url-type';
+
+
+async function getCategoryByOrganizationId(organizationId: Organization['_id'], page: number = 1, limit: number = 10) {
   try {
-    return (
-      await api.get(
-        `v1/api/organizations/${organizationId}/categories?page=${page}&limit=${limit}`,
-      )
-    ).data.payload;
+    return (await api.get(`v1/api/organizations/${organizationId}/categories?page=${page}&limit=${limit}`)).data
+      .payload;
   } catch (e: any) {
     console.log(e.message);
   }
@@ -28,9 +22,7 @@ async function searchCategoryByOrganizationId(
 ) {
   try {
     return (
-      await api.get(
-        `v1/api/organizations/${organizationId}/categories/search?page=${page}&limit=${limit}&q=${query}`,
-      )
+      await api.get(`v1/api/organizations/${organizationId}/categories/search?page=${page}&limit=${limit}&q=${query}`)
     ).data.payload;
   } catch (e: any) {
     console.log(e.message);
@@ -42,12 +34,7 @@ async function addUrlToCategories(
   payload: { url: Url['_id']; categories: Category['_id'][] },
 ) {
   try {
-    return (
-      await api.patch(
-        `/v1/api/organizations/${organizationId}/categories/add-url`,
-        payload,
-      )
-    ).data.payload;
+    return (await api.patch(`/v1/api/organizations/${organizationId}/categories/add-url`, payload)).data.payload;
   } catch (e: any) {
     console.log(e.message);
   }
