+++ src/types/Organization.ts
@@ -1,9 +1,7 @@
-type Organization = {
+export type Organization = {
   _id: string;
   longName: string;
   shortName: string;
   managers: string[];
   domains: string[];
 };
-
-export default Organization;
