+++ src/types/index.ts
@@ -6,3 +6,4 @@ export * from './SidebarItem';
 export * from './SortOption';
 export * from './Tool';
 export * from './NavItem';
+export * from './Organization';
